requirements.txt
python-telegram-bot
requests
schedule
# Se tiver mais dependências, adicione uma por linha
